Refactor: setStyle.js 생성 통한 중복 스타일 양식 집합
